Pin rhai dependency.
Rhai just released a 1.10 version that adds deprecation notices, and changes the way variables get registered. This causes the router to fail to compile, which is seen through scaffold because the router disallows warnings.

This PR works around it by pinning rhai to 1.9.1, until we update our use of the crate in the router.